requirements.txt
greenlet==3.2.3
mysql-connector-python==9.3.0
numpy==2.2.6
pandas==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
